requirements.txt
discord.py
aiosqlite
Pillow
quart
aiohttp
python-dotenv
